Moved command transformer and wrapper to more proper package.
